# 4.0.1 - timezone bug fixes - performance improves - doc updates