--- a/.gitignore
+++ b/.gitignore
@@ -130,3 +130,5 @@ dist
 .pnp.*
 yarn.lock
 package-lock.json
+.DS_Store
+CLAUDE.md