--- a/package.json
+++ b/package.json
@@ -8,6 +8,7 @@
 		"lint": "npx biome lint",
 		"format": "npx biome format",
 		"husky": "npx husky",
+		"benchmark2": "node benchmark2.js",
 		"benchmark": "node benchmark.js"
 	},
 	"author": {