… README.md
… docs/ADVANCED-USAGE.md
… docs/PERFORMANCE.md
… docs/TESTING-GUIDE.md
… index.js